clean up repo + structure
